Added abstract class to add return type
